+++ projects/ngx-forms-inputs/src/inputs/text-input/text-input.component.ts
@@ -1,10 +1,10 @@
 import {ControlContainer, FormsModule} from "@angular/forms";
 import {Component, Host, Optional, SkipSelf} from "@angular/core";
 import {harmonicaAnimation} from "@consensus-labs/ngx-tools";
-import { BaseInputComponent, FormScopeService } from "@consensus-labs/ngx-forms";
-import {MatInputModule} from "@angular/material/input";
+import {BaseInputComponent, FormScopeService} from "@consensus-labs/ngx-forms";
 import {AsyncPipe, NgIf} from "@angular/common";
-import {MatTooltipModule} from "@angular/material/tooltip";
+import {MatLegacyTooltipModule} from "@angular/material/legacy-tooltip";
+import {MatLegacyInputModule} from "@angular/material/legacy-input";
 
 @Component({
   selector: 'form-text-input',
@@ -12,9 +12,9 @@ import {MatTooltipModule} from "@angular/material/tooltip";
   styleUrls: ['./text-input.component.scss'],
   animations: [harmonicaAnimation()],
   imports: [
-    MatInputModule,
+    MatLegacyInputModule,
     NgIf,
-    MatTooltipModule,
+    MatLegacyTooltipModule,
     AsyncPipe,
     FormsModule
   ],